+++ familyhelper-note-impl/src/main/java/com/dwarfeng/familyhelper/note/impl/dao/preset/PonbPresetCriteriaMaker.java
@@ -9,7 +9,10 @@ import org.hibernate.criterion.Restrictions;
 import org.springframework.stereotype.Component;
 
 import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Component
 public class PonbPresetCriteriaMaker implements PresetCriteriaMaker {
@@ -28,6 +31,9 @@ public class PonbPresetCriteriaMaker implements PresetCriteriaMaker {
                 break;
             case PonbMaintainService.CHILD_FOR_USER_PERMISSION_LEVEL_EQUALS:
                 childForUserPermissionLevelEquals(detachedCriteria, objects);
+                break;
+            case PonbMaintainService.CHILD_FOR_NOTE_BOOK_SET:
+                childForNoteBookSet(detachedCriteria, objects);
                 break;
             default:
                 throw new IllegalArgumentException("无法识别的预设: " + s);
@@ -95,6 +101,24 @@ public class PonbPresetCriteriaMaker implements PresetCriteriaMaker {
             }
             int permissionLevel = (int) objects[1];
             detachedCriteria.add(Restrictions.eq("permissionLevel", permissionLevel));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void childForNoteBookSet(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            Collection<LongIdKey> noteBookKeys;
+            if (Objects.isNull(objects[0])) {
+                noteBookKeys = Collections.emptySet();
+            } else {
+                @SuppressWarnings({"UnnecessaryLocalVariable", "unchecked"})
+                Collection<LongIdKey> dejaVu = (Collection<LongIdKey>) objects[0];
+                noteBookKeys = dejaVu;
+            }
+            Collection<Long> noteBookLongIds = noteBookKeys.stream().map(LongIdKey::getLongId)
+                    .collect(Collectors.toSet());
+            detachedCriteria.add(Restrictions.in("noteBookLongId", noteBookLongIds));
         } catch (Exception e) {
             throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
         }